new module for clipboard parsing
